Improved leveling
We were handling the leveling in a loop, but if you crank up the EXP rate and level a couple thousand times in a loop, the client, weirdly enough, starts to struggle =D" Modified it a little to only send the level up updates once and added a max level check.

+++ src/BarracksServer/Database/Character.cs
@@ -192,7 +192,7 @@ namespace Melia.Barracks.Database
 		/// <summary>
 		/// Returns the max EXP for the character's current level.
 		/// </summary>
-		public int MaxExp => BarracksServer.Instance.Data.ExpDb.GetNextExp(this.Level);
+		public long MaxExp => BarracksServer.Instance.Data.ExpDb.GetNextExp(this.Level);
 
 		/// <summary>
 		/// Creates a new character with default values.
